requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
websockets
pydantic
python-jose[cryptography]
passlib
bcrypt
python-dotenv
python-multipart
email-validator
pytest
pytest-asyncio
aiosqlite